backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
langchain==1.2.13
langchain-core==1.2.22
langchain-google-genai==4.2.1
google-genai==1.68.0
pydantic==2.12.5
python-dotenv==1.2.2
